Implement yarn fn params for multiple data types
